# api/openapi.yaml
openapi: 3.1.1
info:
  title: Cooking Assistant API
  version: 1.0.0
  description: API for recipe management and cooking assistance
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /users/register:
    post:
      tags: [Users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
        "400":
          description: Invalid request body (e.g. blank username or password, username contains invalid characters)
        "409":
          description: Username already taken
  /users/login:
    post:
      tags: [Users]
      summary: Login and receive a JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid username or password
  /users/logout:
    post:
      tags: [Users]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
        "401":
          description: Missing or invalid token
  /users/profile:
    get:
      tags: [Users]
      summary: Get current user profile and preferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile and preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing or invalid token
    put:
      tags: [Users]
      summary: Update user profile and preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Profile and preferences updated
        "400":
          description: Invalid request body (e.g. blank password, username contains invalid characters)
        "401":
          description: Missing or invalid token
        "409":
          description: Username already taken
    delete:
      tags: [Users]
      summary: Delete current user account and all associated recipes (cascade)
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User account and all associated data deleted
        "401":
          description: Missing or invalid token

  # -------------------------
  # RECIPES
  # -------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------
  # SCHEMAS
  # -------------------------
  schemas:
    AuthResponse:
      type: object
      required: [token]
      additionalProperties: false
      properties:
        token:
          type: string
          minLength: 1
          description: JWT bearer token to include in subsequent requests
    RegisterRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          description: Alphanumeric, underscores, and hyphens only
        password:
          type: string
          minLength: 4
          maxLength: 128
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    UserProfile:
      type: object
      required: [username, preferences]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
        preferences:
          $ref: "#/components/schemas/UserPreferences"
    UserProfileUpdate:
      type: object
      additionalProperties: false
      description: At least one field must be provided
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
        password:
          type: string
          minLength: 4
          maxLength: 128
        preferences:
          $ref: "#/components/schemas/UserPreferences"
    UserPreferences:
      type: object
      additionalProperties: false
      properties:
        diet:
          type: string
          description: Dietary restriction or style (e.g. vegan, keto)
        allergies:
          type: array
          items:
            type: string
            minLength: 1
          description: List of ingredients the user is allergic to
        aboutMe:
          type: array
          items:
            type: string
            minLength: 1
          description: Free-form user context provided to the AI
...